requirements.txt
customtkinter~=5.2.0
Pillow~=9.5.0
googletrans~=4.0.0rc1